requirements.txt
flask==1.1.2
python-dotenv~=0.10.1
flask-restplus==0.13.0
sqlalchemy>=1.2.0b3
flask-sqlalchemy==2.4.1
Flask-Migrate==2.1.1
requests==2.23.0
flask-cors==3.0.7
flask-bcrypt==0.7.1
sentry-sdk[flask]==0.14.3
Werkzeug==0.16.1
pymysql==0.9.3